iot/requirements.txt
absl-py==0.8.1
asgiref==3.4.1
astor==0.8.0
attrs==21.2.0
bidict==0.21.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
dataclasses==0.8
Django==3.2.6
django-socketio==0.2
dlib @ file:///./dlib-19.8.1-cp36-cp36m-win_amd64.whl
gast==0.2.2
gevent==21.8.0
gevent-socketio==0.3.6
gevent-websocket==0.10.1
google-pasta==0.1.7
greenlet==1.1.1
grpcio==1.24.1
h5py==2.10.0
idna==3.2
imutils==0.5.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mediapipe==0.8.3
mock==3.0.5
mtcnn==0.0.9
numpy==1.19.3
onnx==1.6.0
onnx-tf==1.3.0
onnxruntime==0.5.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
Pillow==8.3.1
protobuf==3.17.3
pycparser==2.20
python-engineio==4.2.1
python-engineio-client==1.1
python-socketio==5.4.0
python-socketio-client==1.1
pytz==2021.1
PyYAML==5.1.2
requests==2.26.0
six==1.12.0
sphinx-me==0.3
sqlparse==0.4.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typing-extensions==3.7.4
urllib3==1.26.7
websocket-client==1.2.1
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
zope.event==4.5.0
zope.interface==5.4.0
